Split ExecutionOptions into OptionsParser and OptionsParser.SuppliedOptions
