+++ src/main/java/net/hydromatic/sqllogictest/SltTestFile.java
@@ -248,10 +248,10 @@ public class SltTestFile {
   }
 
   /**
-   * Parse the contents of the file.
-   * @param options      Options which guide the parsing.
+   * Parse the contents of the test file.
+   * @param options  Options guiding execution.
    */
-  public void parse(ExecutionOptions options) throws IOException {
+  public void parse(OptionsParser.SuppliedOptions options) throws IOException {
     PostgresPolicy policy = new PostgresPolicy();
 
     String line;
@@ -310,7 +310,8 @@ public class SltTestFile {
     }
   }
 
-  private void add(ISqlTestOperation operation, ExecutionOptions options) {
+  private void add(ISqlTestOperation operation,
+                   OptionsParser.SuppliedOptions options) {
     options.message("Operation added " + operation.toString(), 2);
     this.fileContents.add(operation);
     if (operation.is(SqlTestQuery.class)) {
