+++ src/main/java/net/hydromatic/sqllogictest/executors/JdbcExecutor.java
@@ -22,8 +22,8 @@
  */
 package net.hydromatic.sqllogictest.executors;
 
-import net.hydromatic.sqllogictest.ExecutionOptions;
 import net.hydromatic.sqllogictest.ISqlTestOperation;
+import net.hydromatic.sqllogictest.OptionsParser;
 import net.hydromatic.sqllogictest.SltSqlStatement;
 import net.hydromatic.sqllogictest.SltTestFile;
 import net.hydromatic.sqllogictest.SqlTestQuery;
@@ -140,8 +140,8 @@ public abstract class JdbcExecutor extends SqlSltTestExecutor {
    * @param username Name of database user.
    * @param password Password of database user.
    */
-  public JdbcExecutor(ExecutionOptions options, String dbUrl, String username,
-      String password) {
+  public JdbcExecutor(OptionsParser.SuppliedOptions options, String dbUrl,
+                      String username, String password) {
     super(options);
     this.dbUrl = dbUrl;
     this.username = username;
@@ -427,7 +427,7 @@ public abstract class JdbcExecutor extends SqlSltTestExecutor {
    * @return        The statistics describing the tests executed.
    */
   @Override public TestStatistics execute(SltTestFile file,
-      ExecutionOptions options)
+      OptionsParser.SuppliedOptions options)
       throws SQLException, NoSuchAlgorithmException {
     this.startTest();
     this.establishConnection();
